Add optional letters on the bottom of buttons
Introduce the new showLetters field for PinCode, wich defaults to false.

Switch the old IconButton widgets with MaterialButton for ink-splash.

Make the shape of MaterialButton rounded with a bit of padding.

Add the spacer (SizedBox) and the Text for letters as List (workaround).
This could be changed to use spread-operators in the future.
At the moment spread-operators are not in stable channel.

Swap the empty IconButton on the "bottom-left" with empty Expanded.

Add the new introduced option to the example.

Resolves #7